removed num_rounds argument from snpe_a
